+++ Cudheart/Testing/python/test.py
@@ -1,7 +1,11 @@
 from util import *
 
-test_name = "IO::save<int>(string, Vector<StringType*>*, char)"
-res = np.fromfile('savedArray.txt', sep='x', dtype=int)
+test_name = "MatrixOps::vander(Vector<int>, int, bool)"
+vec = np.arange(0, 6, 1, dtype=int)
+res = np.vander(vec, 4, True)
 
-out = [11, 21, 31, 41]
+out = [[1, 1, 1, 1], [1, 1, 5, 4], [3, 2, 1, 0], [25, 16, 9, 4], [1, 0, 125, 64], [27, 8, 1, 0]]
+print(vec)
+print(res)
+print(out)
 print(check(test_name, res, out))